uwsd: avoid truncating Content-Length on 32-bit targets
send_file() formatted the file size through a size_t cast, truncating the
64-bit off_t to 32 bits on 32-bit builds and advertising the wrong length
for files of 4 GiB or more. Format the size as uintmax_t; the buffer is
already sized for a 64-bit value.

+++ http.c
@@ -1360,7 +1360,7 @@ send_file(uwsd_client_context_t *cl, uint16_t code, const char *reason, const ch
 	    uwsd_file_if_modified_since(cl, s) &&
 	    uwsd_file_if_range(cl, s))
 	{
-		snprintf(szbuf, sizeof(szbuf), "%zu", (size_t)s->st_size);
+		snprintf(szbuf, sizeof(szbuf), "%ju", (uintmax_t)s->st_size);
 
 		if (!type || !*type)
 			type = uwsd_file_mime_lookup(path);
